gdb/testsuite: disable gdb.cp/call-method-register.exp when not using gcc
The test gdb.cp/call-method-register.exp assumes that the class will be
placed on a register. However, this keyword has been deprecated since
C++11, and Clang, for instance, does not feel the need to follow it.
Since this test is not usable without this working, this commit marks
this test as untested.

Approved-by: Tom Tromey <tom@tromey.com>