sim/ppc: fix warnings related to printf format strings
This commit is a follow on to:

  commit 173d858e0bec2f5ebd83e935f7c600ce625ef24e
  Date:   Tue Oct 11 15:02:08 2022 +0100

      sim/ppc: fixes for arguments to printf style functions

where commit 173d858e0bec addressed issues with printf format
arguments that were causing the compiler to give an error, this commit
addresses issues that caused the compiler to emit a warning.

This commit is mostly either changing the format string to match the
argument, or in some cases, excess, unused arguments are removed.